+++ conf/script/src/build_system/cmd/setup/meson_utils.py
@@ -11,7 +11,7 @@ import utils.cli.hidden_prints
 import utils.cmd_integrity
 import utils.error.cls_def
 from build_system.cmd.hierarchy.consts import BUILD_SYSTEM_NAME
-from build_system.cmd.setup.cli.cli_print_meson_cmd import print_meson_cmd, print_meson_main_file
+from build_system.cmd.setup.cli.meson import print_meson_cmd, print_meson_main_file
 from build_system.cmd.setup.cli.cli_print_target_info import print_target_info
 from build_system.cmd.setup.meson_machine_file_cli_args import generate_meson_machine_files_cli_args
